project4/requirements.txt
streamlit
requests
python-dotenv
langchain-openai
langchain-core
